Add make-keyed-dynamic-variable primitive
The binder does not tail call the second argument since a new
continuation is created which simply stores the value and which passes
through the result of calling the second argument. It's possible to make
the call a tail call by preserving dynamic variables across tail calls'
continuations, but for now this is not done.