Reimplement $remote-eval in RPython2 interpreter
Note that this function modifies the parent continuation's debug into to
be the given expression. This makes the function useful for fexprs to
give a more user-friendly stack trace upon failure.

This modification should not cause any visible changes to any program.
When continuation guarding is implemented, check that replacing eval
with $remote-eval does not skip a guard.