Remove environment versioning
It does not work with accessing bindings of a virtual environment. An
example of non-local variables lookups is code inside a $let block. The
variables outside the $let are usually arguments of a surrounding $vau
or $lambda, not constants.

Note that this does add a few more checks to the compiled loop, though
all checks should have been optimized away from the loop-peeling
optimization. Further investigation is needed since some functions are
guarded before the loop.